+++ R/raw.getType.R
@@ -2,10 +2,10 @@
 #' @returns string with type, such as "XRD", "AFM", etc.
 #' @export
 raw.getType <- function(rawData, ID) {
-  m = which(rawData$ID == ID)
+  m = which(rawData$df$ID == ID)
   rawType = ""
   if(length(m)==1) {
-    rawType = rawData$type[m]
+    rawType = rawData$df$type[m]
   }
   rawType
 }